Allow each scene capture sensor have its own serializer
