+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Sensor/SceneCaptureSensor.cpp
@@ -93,41 +93,6 @@ float ASceneCaptureSensor::GetFOVAngle() const
   return CaptureComponent2D->FOVAngle;
 }
 
-bool ASceneCaptureSensor::ReadPixels_GameThread(TArray<FColor> &BitMap) const
-{
-  check(IsInGameThread());
-  if (!CaptureRenderTarget)
-  {
-    UE_LOG(LogCarla, Error, TEXT("SceneCaptureCamera: Missing render target"));
-    return false;
-  }
-  FTextureRenderTargetResource *RTResource =
-      CaptureRenderTarget->GameThread_GetRenderTargetResource();
-  if (RTResource == nullptr)
-  {
-    UE_LOG(LogCarla, Error, TEXT("SceneCaptureCamera: Missing render target"));
-    return false;
-  }
-  FReadSurfaceDataFlags ReadPixelFlags(RCM_UNorm);
-  ReadPixelFlags.SetLinearToGamma(true);
-  return RTResource->ReadPixels(BitMap, ReadPixelFlags);
-}
-
-bool ASceneCaptureSensor::SaveCaptureToDisk(const FString &FilePath) const
-{
-  TArray<FColor> OutBMP;
-  if (!ReadPixels_GameThread(OutBMP)) {
-    return false;
-  }
-  for (FColor &color : OutBMP) {
-    color.A = 255u;
-  }
-  const FIntPoint DestSize(GetImageWidth(), GetImageHeight());
-  FString ResultPath;
-  FHighResScreenshotConfig &HighResScreenshotConfig = GetHighResScreenshotConfig();
-  return HighResScreenshotConfig.SaveImage(FilePath, OutBMP, DestSize, &ResultPath);
-}
-
 void ASceneCaptureSensor::PostActorCreated()
 {
   Super::PostActorCreated();
